Conflicts:
	pom.xml
	src/main/java/org/training/kparfiankou/issuetracker/beans/AbstractEntity.java
	src/main/java/org/training/kparfiankou/issuetracker/beans/Build.java
	src/main/java/org/training/kparfiankou/issuetracker/beans/Password.java
	src/main/java/org/training/kparfiankou/issuetracker/beans/Priority.java
	src/main/java/org/training/kparfiankou/issuetracker/beans/Project.java
	src/main/java/org/training/kparfiankou/issuetracker/beans/Resolution.java
	src/main/java/org/training/kparfiankou/issuetracker/beans/SimpleEntiry.java
	src/main/java/org/training/kparfiankou/issuetracker/beans/Status.java
	src/main/java/org/training/kparfiankou/issuetracker/beans/Type.java
	src/main/java/org/training/kparfiankou/issuetracker/beans/User.java
	src/main/java/org/training/kparfiankou/issuetracker/controllers/CreateUpdateProjectController.java
	src/main/java/org/training/kparfiankou/issuetracker/controllers/InsertAddingSimpleEntityController.java
	src/main/java/org/training/kparfiankou/issuetracker/controllers/MainController.java
	src/main/java/org/training/kparfiankou/issuetracker/factories/TypeDAOFactory.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/database/ProjectDatabaseDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/hibernate/PriorityHibernateDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/hibernate/ProjectHibernateDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/hibernate/ResolutionHibernateDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/hibernate/StatusHibernateDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/hibernate/TypeHibernateDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/hibernate/UserHibernateDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/xml/IssueXMLDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/xml/PriorityXMLDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/xml/ProjectXMLDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/xml/ResolutionXMLDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/xml/StatusXMLDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/xml/TypeXMLDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/impl/xml/UserXMLDAO.java
	src/main/java/org/training/kparfiankou/issuetracker/interfaces/IStatusDAO.java
	src/main/resources/hibernate.cfg.xml

+++ src/main/java/org/training/kparfiankou/issuetracker/impl/database/TypeDatabaseDAO.java
@@ -162,13 +162,13 @@ public class TypeDatabaseDAO extends AbstractDatabaseDAO implements ITypeDAO {
 	}
 
 	@Override
-	public void removeType(int id) {
+	public void removeType(long id) {
 
 		final int numId = 1;
 
 		try {
 
-			psRemoveType.setInt(numId, id);
+			psRemoveType.setLong(numId, id);
 			psRemoveType.executeUpdate();
 
 		} catch (SQLException e) {
@@ -179,7 +179,7 @@ public class TypeDatabaseDAO extends AbstractDatabaseDAO implements ITypeDAO {
 	}
 
 	@Override
-	public int getMaxIndex() {
+	public long getMaxIndex() {
 
 		int numIdColum = 1;
 		int maxId = 0;
